renamed original startPGame to startSnongGame (obsolete, to be removed soon). Split startSGame into gameSetup (interfaces/functions to be used for both Pong/Snek) and startSGame (functions to be used just for Snek). Duplicated the split startSGame into startPGame, started to make adjustments for it to be Pong
