+++ srcs/frontend/src/pages/startSGame.ts
@@ -5,45 +5,34 @@ import { setupErrorPages } from './errorPages';
 import DOMPurify from 'dompurify';
 import { connectFunc, requestBody } from '../script/connections';
 import { setupSnek } from './snek';
+import { AuthState } from '../script/gameSetup'
+import { FormToggleListener, updateStartGameButton, setupGuestAliasLocking, setupLoginValidation, newPlayersButton } from '../script/gameSetup'
 
 import "../styles/snek.css"
 
-interface AuthState {
-    isAuthenticated: boolean;
-    isGuestLocked: boolean;
-    guestAlias: string;
-    userAlias: string;
-    userUuid?: string; // Added UUID for authenticated users
-}
-
 interface PlayerStats {
-    alias: string;
-    matches: number;
-    wins: number;
-    losses: number;
-    winrate: number;
-    avg_score: number;
-    highest_score: number;
-}
-
-interface UserData {
-    uuid: string;
-    alias: string;
+	alias: string;
+	matches: number;
+	wins: number;
+	losses: number;
+	winrate: number;
+	avg_score: number;
+	highest_score: number;
 }
 
 interface GameEndPayload {
-    p2_alias: string;
-    p2_uuid?: string;
-    winner_id: number;
-    p1_score: number;
-    p2_score: number;
+	p2_alias: string;
+	p2_uuid?: string;
+	winner_id: number;
+	p1_score: number;
+	p2_score: number;
 }
 
 const authState: AuthState = {
-    isAuthenticated: false,
-    isGuestLocked: false,
-    guestAlias: "",
-    userAlias: ""
+	isAuthenticated: false,
+	isGuestLocked: false,
+	guestAlias: "",
+	userAlias: ""
 };
 
 export function setupStartSGame() {
@@ -150,14 +139,21 @@ export function setupStartSGame() {
         }
         const container = document.getElementById('gameContainer') as HTMLElement;
         if (container) {
-            preGameScreen(container).then((app: Application) => {
+            preGameScreen(container).then(async (app: Application) => {
 
-                setupGuestAliasLocking();
-                FormToggleListener();
-                setupLoginValidation(app);
-                updateStartGameButton();
+                setupGuestAliasLocking(authState);
+                FormToggleListener(authState);
+                setupLoginValidation(app, authState);
+				if (authState.isAuthenticated)
+				{
+					// Fetch and display player2 stats
+					const player2Stats = await fetchPlayer2Stats(authState.userAlias);
+					if (player2Stats)
+						updatePlayer2StatsDisplay(player2Stats);
+				}
+                updateStartGameButton(authState);
                 startGameListeners(app);
-                newPlayersButton(app);
+                newPlayersButton(app, authState);
             }).catch((error) => {
                 console.error("Error setting up the game:", error);
                 window.history.pushState({}, '', '/errorPages');
@@ -171,143 +167,7 @@ export function setupStartSGame() {
 
 }
 
-// Prevents the toggle from being used if user is logged in / guest is locked in
-function FormToggleListener() {
-    const toggle = document.getElementById("authToggle") as HTMLInputElement;
-    if (!toggle) {
-        console.error("Auth toggle not found");
-        return;
-    }
-
-    toggle.addEventListener('change', () => {
-        if (authState.isGuestLocked || authState.isAuthenticated) {
-            toggle.checked = !toggle.checked;
-            return;
-        }
-        updateFormToggle();
-    });
-}
-
-// updates the state of the form toggle
-function updateFormToggle() {
-    const toggle = document.getElementById("authToggle") as HTMLInputElement;
-    const toggleBackground = document.getElementById("toggleBackground");
-    const toggleCircle = document.getElementById("toggleCircle");
-    const guestForm = document.getElementById("GuestAliasform") as HTMLFormElement;
-    const loginForm = document.getElementById("LoginForm") as HTMLFormElement;
-
-    if (!toggle || !toggleBackground || !toggleCircle || !guestForm || !loginForm) {
-        console.error("Form toggle not found");
-        return;
-    }
-    if (toggle.checked) {
-        toggleBackground.classList.add('bg-blue-600');
-        toggleBackground.classList.remove('bg-gray-300');
-        toggleCircle.style.transform = 'translateX(32px)'
-        guestForm.classList.add('hidden');
-        guestForm.classList.remove('flex');
-        loginForm.classList.remove('hidden');
-        loginForm.classList.add('flex');
-    } else {
-        toggleBackground.classList.remove('bg-blue-600');
-        toggleBackground.classList.add('bg-gray-300');
-        toggleCircle.style.transform = 'translateX(0)';
-        loginForm.classList.add('hidden');
-        loginForm.classList.remove('flex');
-        guestForm.classList.remove('hidden');
-        guestForm.classList.add('flex');
-
-        const loginStatus = document.getElementById('loginStatus');
-        if (loginStatus) {
-            loginStatus.classList.add('hidden');
-        }
-        (document.getElementById('loginUsername') as HTMLInputElement).value = '';
-        (document.getElementById('loginPassword') as HTMLInputElement).value = '';
-    }
-}
-
-// enables the start game button and displays the player2 alias
-function updateStartGameButton() {
-    const startGameButton = document.getElementById('startGame') as HTMLButtonElement;
-    const player2InfoElements = document.querySelectorAll('.player2-info');
-    const player2StatsContainer = document.getElementById('player2StatsContainer');
-
-    if (!startGameButton) {
-        console.error("Start game button not found");
-        return;
-    }
-    if (authState.isAuthenticated || authState.isGuestLocked) {
-        startGameButton.disabled = false;
-        startGameButton.classList.remove('bg-gray-500', 'cursor-not-allowed', 'opacity-50');
-        startGameButton.classList.add('bg-blue-500', 'hover:bg-blue-700', 'text-white');
-
-        const displayName = authState.isAuthenticated ? authState.userAlias : authState.guestAlias;
-        player2InfoElements.forEach(element => {
-            element.textContent = displayName;
-        });
-
-        // Show player2 stats if authenticated
-        if (player2StatsContainer) {
-            if (authState.isAuthenticated) {
-                player2StatsContainer.classList.remove('hidden');
-            } else {
-                player2StatsContainer.classList.add('hidden');
-            }
-        }
-    } else {
-        startGameButton.disabled = true;
-        startGameButton.classList.add('bg-gray-500', 'cursor-not-allowed', 'opacity-50');
-        startGameButton.classList.remove('bg-blue-500', 'hover:bg-blue-700', 'text-white');
-        player2InfoElements.forEach(element => {
-            element.textContent = "";
-        });
-
-        // Hide player2 stats
-        if (player2StatsContainer) {
-            player2StatsContainer.classList.add('hidden');
-        }
-    }
-}
-
-// validates and locks/unlocks the guest alias
-function setupGuestAliasLocking() {
-    const guestInputField = document.getElementById("guestAliasInput") as HTMLInputElement;
-    const lockInButton = document.getElementById("lockInGuest") as HTMLButtonElement;
-    const changeButton = document.getElementById("changeGuestAlias") as HTMLButtonElement;
-
-    if (!guestInputField || !lockInButton || !changeButton) {
-        console.error("Guest alias input or buttons not found");
-        return;
-    }
-
-    lockInButton.addEventListener('click', () => {
-        const rawInput = guestInputField.value.trim();
-        const sanitizedInput = rawInput.replace(/[^a-zA-Z0-9]/g, '');
-        const guestInput = DOMPurify.sanitize(sanitizedInput);
-
-        if (guestInput.length < 3) {
-            alert("Guest alias must be at least 3 characters long.");
-            return;
-        }
-        guestInputField.disabled = true;
-        lockInButton.classList.add('hidden');
-        changeButton.classList.remove('hidden');
-
-        authState.isGuestLocked = true;
-        authState.guestAlias = "(guest) " + guestInput;
-        updateStartGameButton();
-    });
-
-    changeButton.addEventListener('click', () => {
-        guestInputField.disabled = false;
-        changeButton.classList.add('hidden');
-        lockInButton.classList.remove('hidden');
-        authState.isGuestLocked = false;
-        updateStartGameButton();
-    });
-}
-
-// Function to update player2 stats display
+// Function to update player2 stats display (for Snek)
 function updatePlayer2StatsDisplay(stats: PlayerStats) {
     const matches = document.getElementById('p2-matches');
     const wins = document.getElementById('p2-wins');
@@ -324,7 +184,7 @@ function updatePlayer2StatsDisplay(stats: PlayerStats) {
     if (highestScore) highestScore.textContent = stats.highest_score.toString();
 }
 
-// Function to fetch player2 stats
+// Function to fetch player2 stats (for Snek)
 async function fetchPlayer2Stats(alias: string): Promise<PlayerStats | null> {
     try {
         const sanitizedAlias = DOMPurify.sanitize(alias);
@@ -345,118 +205,7 @@ async function fetchPlayer2Stats(alias: string): Promise<PlayerStats | null> {
     }
 }
 
-// logs the user in
-function setupLoginValidation(app: Application) {
-    const loginButton = document.getElementById('loginButton');
-    const logoutButton = document.getElementById('logoutButton');
-    const usernameInput = document.getElementById('loginUsername') as HTMLInputElement;
-    const passwordInput = document.getElementById('loginPassword') as HTMLInputElement;
-    const loginStatus = document.getElementById('loginStatus');
-
-    if (!loginButton || !logoutButton || !usernameInput || !passwordInput || !loginStatus) {
-        console.error("Login elements not found");
-        return;
-    }
-
-    loginButton.addEventListener('click', async () => {
-        const usernameIn = usernameInput.value.trim().replace(/[^a-zA-Z0-9]/g, '');
-        const passwordIn = passwordInput.value.trim().replace(/[^a-zA-Z0-9]/g, '');
-        const username = DOMPurify.sanitize(usernameIn);
-        const password = DOMPurify.sanitize(passwordIn);
-        if (!username || username.length < 3) {
-            showLoginStatus(loginStatus, "Username must be at least 3 characters long", false);
-            return;
-        }
-        if (!password || password.length < 6) {
-            showLoginStatus(loginStatus, "Password must be at least 6 characters long", false);
-            return;
-        }
-        try {
-            const userData = await validateLogin(username, password);
-            if (userData) {
-                showLoginStatus(loginStatus, "Login successful!", true);
-                authState.isAuthenticated = true;
-                authState.userAlias = userData.alias;
-                authState.userUuid = userData.uuid; // Store UUID for game results
-
-                usernameInput.disabled = true;
-                passwordInput.disabled = true;
-                loginButton.classList.add('hidden');
-                logoutButton.classList.remove('hidden');
-
-                // Fetch and display player2 stats
-                const player2Stats = await fetchPlayer2Stats(userData.alias);
-                if (player2Stats) {
-                    updatePlayer2StatsDisplay(player2Stats);
-                }
-
-                updateStartGameButton();
-                updateFormToggle();
-            } else {
-                showLoginStatus(loginStatus, "Invalid username or password", false);
-                authState.isAuthenticated = false;
-                updateStartGameButton();
-                updateFormToggle();
-            }
-        } catch (error) {
-            showLoginStatus(loginStatus, "Error during login. Please try again.", false);
-            console.error("Login error:", error);
-            authState.isAuthenticated = false;
-            updateStartGameButton();
-            updateFormToggle();
-        }
-    });
-
-    logoutButton.addEventListener('click', () => {
-        authState.isAuthenticated = false;
-        authState.userAlias = "";
-        authState.userUuid = undefined;
-
-        usernameInput.disabled = false;
-        passwordInput.disabled = false;
-        usernameInput.value = "";
-        passwordInput.value = "";
-        loginButton.classList.remove('hidden');
-        logoutButton.classList.add('hidden');
-        loginStatus.classList.add('hidden');
-
-        resetGame(app);
-        updateStartGameButton();
-        updateFormToggle();
-    });
-}
-
-// displays output of login attempt
-function showLoginStatus(statusElement: HTMLElement, message: string, isSuccess: boolean) {
-    statusElement.textContent = message;
-    statusElement.classList.remove('hidden', 'text-green-500', 'text-red-500');
-    statusElement.classList.add(isSuccess ? 'text-green-500' : 'text-red-500');
-}
-
-// calls login API
-async function validateLogin(username: string, password: string): Promise<UserData | null> {
-    console.log(`Attempting to login with username: ${username}`);
-
-    try {
-        const response = await connectFunc(
-            "/user/game/login",
-            requestBody("POST", JSON.stringify({ username, password }), "application/json")
-        );
-
-        if (response.ok) {
-            const userData: UserData = await response.json();
-            return userData;
-        } else {
-            console.error(`Login failed: ${response.status}`);
-            return null;
-        }
-    } catch (error) {
-        console.error("Login error:", error);
-        return null;
-    }
-}
-
-// Record game results
+// Record game results (for Snek)
 async function recordGameResults(gameData: gameEndData): Promise<boolean> {
     try {
         // Prepare the payload for the API
@@ -493,7 +242,7 @@ async function recordGameResults(gameData: gameEndData): Promise<boolean> {
     }
 }
 
-// starts the listeners for the game button
+// starts the listeners for the game button (for Snek)
 async function startGameListeners(app: Application): Promise<void> {
     const startGameButton = document.getElementById('startGame') as HTMLButtonElement;
     const restartGameButton = document.getElementById('restartGame');
@@ -572,85 +321,5 @@ async function startGameListeners(app: Application): Promise<void> {
 
         replayButtons.classList.remove('hidden');
         replayButtons.classList.add('flex');
-    });
-}
-
-// logic for resetting the game ( newplayer button )
-// logic for resetting the game ( newplayer button )
-function newPlayersButton(app: Application) {
-    const newGameButton = document.getElementById('newGame');
-    if (!newGameButton) {
-        console.error("New game button not found");
-        return;
-    }
-
-    newGameButton.addEventListener('click', () => {
-        window.history.pushState({}, '', '/home');
-
-        // Reset auth state
-        authState.isAuthenticated = false;
-        authState.isGuestLocked = false;
-        authState.guestAlias = "";
-        authState.userAlias = "";
-        authState.userUuid = undefined;
-
-        // Reset guest form
-        const guestInput = document.getElementById("guestAliasInput") as HTMLInputElement;
-        const lockInButton = document.getElementById("lockInGuest") as HTMLButtonElement;
-        const changeButton = document.getElementById("changeGuestAlias") as HTMLButtonElement;
-
-        if (guestInput && lockInButton && changeButton) {
-            guestInput.value = "";
-            guestInput.disabled = false;
-            lockInButton.classList.remove('hidden');
-            changeButton.classList.add('hidden');
-        }
-
-        // Reset login form
-        const usernameInput = document.getElementById("loginUsername") as HTMLInputElement;
-        const passwordInput = document.getElementById("loginPassword") as HTMLInputElement;
-        const loginButton = document.getElementById("loginButton") as HTMLButtonElement;
-        const logoutButton = document.getElementById("logoutButton") as HTMLButtonElement;
-        const loginStatus = document.getElementById("loginStatus");
-
-        if (usernameInput && passwordInput && loginStatus) {
-            usernameInput.value = "";
-            passwordInput.value = "";
-            usernameInput.disabled = false;
-            passwordInput.disabled = false;
-            loginStatus.classList.add('hidden');
-        }
-
-        // Reset login/logout buttons
-        if (loginButton && logoutButton) {
-            loginButton.classList.remove('hidden');
-            logoutButton.classList.add('hidden');
-        }
-
-        // Reset toggle to guest mode
-        const toggle = document.getElementById("authToggle") as HTMLInputElement;
-        if (toggle) {
-            toggle.checked = false;
-            updateFormToggle();
-        }
-
-        // Hide player2 stats
-        const player2StatsContainer = document.getElementById('player2StatsContainer');
-        if (player2StatsContainer) {
-            player2StatsContainer.classList.add('hidden');
-        }
-
-        // Hide replay buttons
-        const replayButtons = document.getElementById('replayButtons');
-        if (replayButtons) {
-            replayButtons.classList.add('hidden');
-            replayButtons.classList.remove('flex');
-        }
-
-        // Update start game button state
-        resetGame(app);
-        updateStartGameButton();
-
-        console.log("New players button clicked");
     });
 }
